Add files via upload
Navigation Drawer with Fragments Part 3 - HANDLING MENU ITEM CLICKS - Android Studio Tutorial

In part 3 of the navigation drawer tutorial, we will create 3 fragments and their corresponding layout files. Then we implement the OnNavigationItemSelectedListener interface into our MainActivity and set it as the listener on our NavigationView to handle click events on our menu items. For our 3 main items we will open the corresponding fragment with the getSupportFragmentManager and beginTransaction methods, and for the other 2 items we will simply show a toast message.

https://www.youtube.com/watch?v=bjYstsO1PgI